+++ ricaun.Nuke.PackageBuilder/Components/Revit/IRevitPackageBuilder.cs
@@ -170,8 +170,20 @@ namespace ricaun.Nuke.Components
             {
                 var folder = file.Parent;
                 SignFolder(folder, $"*{project.Name}*");
-                new RevitProjectAddInsBuilder(project, file, Application, ApplicationType, VendorId, VendorDescription)
-                    .Build(file);
+                var builder = new RevitProjectAddInsBuilder(project, file, Application, ApplicationType, VendorId, VendorDescription);
+                if (RevitContext || RevitContextName != null)
+                {
+                    var fileRevitVersion = RevitExtension.GetRevitVersion(file);
+                    const int ManifestSettingsSupportVersion = 2026;
+                    if (fileRevitVersion >= RevitContextVersion && fileRevitVersion >= ManifestSettingsSupportVersion)
+                    {
+                        var manifestSettings = builder.CreateManifestSettings();
+                        manifestSettings.UseRevitContext = true;
+                        manifestSettings.ContextName = RevitContextName;
+                        Serilog.Log.Information($"Create AddIns ManifestSettings.UseRevitContext in Revit {fileRevitVersion} with ContextName '{RevitContextName}'");
+                    }
+                }
+                builder.Build(file);
             });
         }