+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.11.3] / 2026-04-08
+### Features
+- Support `IHazRevitPackageBuilder.RevitContext` to enable `UseRevitContext` in the addin build.
+### Updated
+- Add `IHazRevitPackageBuilder.RevitContext` to enable `ManifestSettings.UseRevitContext` in the addin build.
+- Add `IHazRevitPackageBuilder.RevitContextName` to set the `ManifestSettings.ContextName` in the addin build.
+- Add `IHazRevitPackageBuilder.RevitContextVersion` to set the lowest version that support the `ManifestSettings` in the addin build.
+
 ## [1.11.2] / 2026-03-22 - 2026-03-30
 ### Features
 - Update `ricaun.Nuke` to `1.11.2` to support `.sln` and/or `.slnx` in the `Solution` in `net10.0`
@@ -390,6 +398,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - First Release
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.11.3]: ../../compare/1.11.2...1.11.3
 [1.11.2]: ../../compare/1.11.1...1.11.2
 [1.11.1]: ../../compare/1.11.0...1.11.1
 [1.11.0]: ../../compare/1.10.0...1.11.0